Adicionado ano letivo ao obter listagem
