+++ src/SME.SGP.Aplicacao/Consultas/ConsultasAtribuicaoCJ.cs
@@ -15,13 +15,11 @@ namespace SME.SGP.Aplicacao
         private readonly IRepositorioAtribuicaoCJ repositorioAtribuicaoCJ;
         private readonly IRepositorioComponenteCurricular repositorioComponenteCurricular;
 
-        private readonly IServicoEol servicoEOL;
 
         public ConsultasAtribuicaoCJ(IRepositorioAtribuicaoCJ repositorioAtribuicaoCJ, IRepositorioComponenteCurricular repositorioComponenteCurricular, IServicoEol servicoEOL)
         {
             this.repositorioAtribuicaoCJ = repositorioAtribuicaoCJ ?? throw new ArgumentNullException(nameof(repositorioAtribuicaoCJ));
             this.repositorioComponenteCurricular = repositorioComponenteCurricular ?? throw new ArgumentNullException(nameof(repositorioComponenteCurricular));
-            this.servicoEOL = servicoEOL ?? throw new ArgumentNullException(nameof(servicoEOL));
         }
 
         public async Task<IEnumerable<AtribuicaoCJListaRetornoDto>> Listar(AtribuicaoCJListaFiltroDto filtroDto)
@@ -32,52 +30,6 @@ namespace SME.SGP.Aplicacao
             if (listaRetorno.Any())
                 return await TransformaEntidadesEmDtosListaRetorno(listaRetorno);
             else return null;
-        }
-
-        public async Task<AtribuicaoCJTitularesRetornoDto> ObterProfessoresTitularesECjs(string ueId, string turmaId,
-            string professorRf, Modalidade modalidadeId)
-        {
-            IEnumerable<ProfessorTitularDisciplinaEol> professoresTitularesDisciplinasEol = await servicoEOL.ObterProfessoresTitularesDisciplinas(turmaId, professorRf);
-
-            var listaAtribuicoes = await repositorioAtribuicaoCJ.ObterPorFiltros(modalidadeId, turmaId, ueId, 0,
-                professorRf, string.Empty, null);
-
-            if (professoresTitularesDisciplinasEol!=null && professoresTitularesDisciplinasEol.Any())
-                return TransformaEntidadesEmDtosAtribuicoesProfessoresRetorno(listaAtribuicoes, professoresTitularesDisciplinasEol);
-            else return null;
-        }
-
-        private AtribuicaoCJTitularesRetornoDto TransformaEntidadesEmDtosAtribuicoesProfessoresRetorno(IEnumerable<AtribuicaoCJ> listaAtribuicoes, IEnumerable<ProfessorTitularDisciplinaEol> professoresTitularesDisciplinasEol)
-        {
-            var listaRetorno = new AtribuicaoCJTitularesRetornoDto();
-
-            foreach (var disciplinaProfessorTitular in professoresTitularesDisciplinasEol)
-            {
-                var atribuicao = listaAtribuicoes.FirstOrDefault(b => b.DisciplinaId == disciplinaProfessorTitular.DisciplinaId);
-
-                listaRetorno.Itens.Add(new AtribuicaoCJTitularesRetornoItemDto()
-                {
-                    Disciplina = disciplinaProfessorTitular.DisciplinaNome,
-                    DisciplinaId = disciplinaProfessorTitular.DisciplinaId,
-                    ProfessorTitular = disciplinaProfessorTitular.ProfessorNome,
-                    ProfessorTitularRf = disciplinaProfessorTitular.ProfessorRf,
-                    Substituir = atribuicao != null && atribuicao.Substituir
-                });
-            }
-
-            if (listaAtribuicoes.Any())
-            {
-                var ultimoRegistroAlterado = listaAtribuicoes
-                    .OrderBy(b => b.AlteradoEm)
-                    .ThenBy(b => b.CriadoEm).FirstOrDefault();
-
-                listaRetorno.CriadoEm = ultimoRegistroAlterado.CriadoEm;
-                listaRetorno.CriadoPor = ultimoRegistroAlterado.CriadoPor;
-                listaRetorno.AlteradoEm = ultimoRegistroAlterado.AlteradoEm;
-                listaRetorno.AlteradoPor = ultimoRegistroAlterado.AlteradoPor;
-            }
-
-            return listaRetorno;
         }
 
         private async Task<IEnumerable<AtribuicaoCJListaRetornoDto>> TransformaEntidadesEmDtosListaRetorno(IEnumerable<AtribuicaoCJ> listaDto)
